python-way/requirements.txt
libvirt-python
jinja2
pyyaml